requirements.txt
typer>=0.12.0
rich>=13.7.0
GitPython>=3.1.43
pyperclip>=1.8.2